feat: 예약 기록 세팅
